# packages/api-contract/openapi.json
{
  "components": {
    "schemas": {
...
      "ChangePasswordRequest": {
        "additionalProperties": false,
        "description": "Current proof and policy-compliant replacement accepted from one user.",
        "properties": {
          "current_password": {
            "format": "password",
            "maxLength": 128,
            "minLength": 1,
            "title": "Current Password",
            "type": "string",
            "writeOnly": true
          },
          "new_password": {
            "format": "password",
            "maxLength": 128,
            "minLength": 12,
            "title": "New Password",
            "type": "string",
            "writeOnly": true
          }
        },
        "required": ["current_password", "new_password"],
        "title": "ChangePasswordRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "AccessToken": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Industry Intelligence Platform API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/auth/change-password": {
      "post": {
        "description": "Replace the password, revoke all sessions, then clear browser state.",
        "operationId": "change_password_api_v1_auth_change_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "description": "Stable, non-sensitive error contract returned by the API.",
                  "properties": {
                    "code": {
                      "title": "Code",
                      "type": "string"
                    },
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "trace_id": {
                      "title": "Trace Id",
                      "type": "string"
                    },
                    "type": {
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": ["type", "title", "status", "detail", "code", "trace_id"],
                  "title": "ProblemDetails",
                  "type": "object"
                }
              }
            },
            "description": "Current password is invalid"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "description": "Stable, non-sensitive error contract returned by the API.",
                  "properties": {
                    "code": {
                      "title": "Code",
                      "type": "string"
                    },
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "trace_id": {
                      "title": "Trace Id",
                      "type": "string"
                    },
                    "type": {
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": ["type", "title", "status", "detail", "code", "trace_id"],
                  "title": "ProblemDetails",
                  "type": "object"
                }
              }
            },
            "description": "Invalid authenticated session"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "description": "Stable, non-sensitive error contract returned by the API.",
                  "properties": {
                    "code": {
                      "title": "Code",
                      "type": "string"
                    },
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "trace_id": {
                      "title": "Trace Id",
                      "type": "string"
                    },
                    "type": {
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": ["type", "title", "status", "detail", "code", "trace_id"],
                  "title": "ProblemDetails",
                  "type": "object"
                }
              }
            },
            "description": "Password change request rejected"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "description": "Stable, non-sensitive error contract returned by the API.",
                  "properties": {
                    "code": {
                      "title": "Code",
                      "type": "string"
                    },
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "trace_id": {
                      "title": "Trace Id",
                      "type": "string"
                    },
                    "type": {
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": ["type", "title", "status", "detail", "code", "trace_id"],
                  "title": "ProblemDetails",
                  "type": "object"
                }
              }
            },
            "description": "Password changed concurrently"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "description": "Stable, non-sensitive error contract returned by the API.",
                  "properties": {
                    "code": {
                      "title": "Code",
                      "type": "string"
                    },
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "trace_id": {
                      "title": "Trace Id",
                      "type": "string"
                    },
                    "type": {
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": ["type", "title", "status", "detail", "code", "trace_id"],
                  "title": "ProblemDetails",
                  "type": "object"
                }
              }
            },
            "description": "Request validation failed"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "description": "Stable, non-sensitive error contract returned by the API.",
                  "properties": {
                    "code": {
                      "title": "Code",
                      "type": "string"
                    },
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "trace_id": {
                      "title": "Trace Id",
                      "type": "string"
                    },
                    "type": {
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": ["type", "title", "status", "detail", "code", "trace_id"],
                  "title": "ProblemDetails",
                  "type": "object"
                }
              }
            },
            "description": "Password change rate limit exceeded"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "description": "Stable, non-sensitive error contract returned by the API.",
                  "properties": {
                    "code": {
                      "title": "Code",
                      "type": "string"
                    },
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "trace_id": {
                      "title": "Trace Id",
                      "type": "string"
                    },
                    "type": {
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": ["type", "title", "status", "detail", "code", "trace_id"],
                  "title": "ProblemDetails",
                  "type": "object"
                }
              }
            },
            "description": "Internal server error"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "description": "Stable, non-sensitive error contract returned by the API.",
                  "properties": {
                    "code": {
                      "title": "Code",
                      "type": "string"
                    },
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "trace_id": {
                      "title": "Trace Id",
                      "type": "string"
                    },
                    "type": {
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": ["type", "title", "status", "detail", "code", "trace_id"],
                  "title": "ProblemDetails",
                  "type": "object"
                }
              }
            },
            "description": "Password change temporarily unavailable"
          }
        },
        "security": [
          {
            "AccessToken": []
          }
        ],
        "summary": "Change Password",
        "tags": ["identity"]
      }
    },
...
    "/api/v1/auth/logout": {
      "post": {
        "description": "Revoke one browser family and clear its cookies after commit.",
        "operationId": "logout_api_v1_auth_logout_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "description": "Stable, non-sensitive error contract returned by the API.",
                  "properties": {
                    "code": {
                      "title": "Code",
                      "type": "string"
                    },
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "trace_id": {
                      "title": "Trace Id",
                      "type": "string"
                    },
                    "type": {
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": ["type", "title", "status", "detail", "code", "trace_id"],
                  "title": "ProblemDetails",
                  "type": "object"
                }
              }
            },
            "description": "Invalid logout session"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "description": "Stable, non-sensitive error contract returned by the API.",
                  "properties": {
                    "code": {
                      "title": "Code",
                      "type": "string"
                    },
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "trace_id": {
                      "title": "Trace Id",
                      "type": "string"
                    },
                    "type": {
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": ["type", "title", "status", "detail", "code", "trace_id"],
                  "title": "ProblemDetails",
                  "type": "object"
                }
              }
            },
            "description": "Internal server error"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "description": "Stable, non-sensitive error contract returned by the API.",
                  "properties": {
                    "code": {
                      "title": "Code",
                      "type": "string"
                    },
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "trace_id": {
                      "title": "Trace Id",
                      "type": "string"
                    },
                    "type": {
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": ["type", "title", "status", "detail", "code", "trace_id"],
                  "title": "ProblemDetails",
                  "type": "object"
                }
              }
            },
            "description": "Logout temporarily unavailable"
          }
        },
        "summary": "Logout",
        "tags": ["identity"]
      }
    },
...
  }
}